add data disk to cvmfs server

+++ cvmfs/setup-server.sh
@@ -1,5 +1,16 @@
 #!/bin/bash
 
+if [ ! -d /data0 ] ; then
+  echo "Please add data disk (/data0) before installing cvmfs server"
+  exit 1
+fi
+
+mkdir -pv /data0/{cvmfs,cvmfs-srv}
+mkdir -pv /srv
+
+ln -sf /data0/cvmfs /var/spool/cvmfs
+ln -sf /data0/cvmfs-srv /srv/cvmfs
+
 wget https://ecsft.cern.ch/dist/cvmfs/cvmfs-release/cvmfs-release-latest_all.deb
 sudo dpkg -i cvmfs-release-latest_all.deb
 rm -f cvmfs-release-latest_all.deb
